src/services/ai/RoutePredictionService.ts
import { Route, RouteSegment, RoutePreferences } from '@/types/route/types';
import { WeatherConditions } from '@/types/weather/types';
import { TerrainAnalysisResult } from '@/types/terrain/types';
import { TrafficConditions } from '@/types/traffic/types';
import { LearningSystem } from './LearningSystem';
import { CrossActivityLearningService } from './CrossActivityLearningService';
import { ContextualLearningService } from './ContextualLearningService';

interface PredictionContext {
  weather: WeatherConditions;
  terrain: TerrainAnalysisResult;
  traffic: TrafficConditions;
  timeOfDay: number;
  dayOfWeek: number;
  season: string;
  userHistory?: Route[];
}

interface PredictionResult {
  predictedRoute: Route;
  confidence: number;
  alternativeRoutes: Route[];
  factors: {
    weatherImpact: number;
    terrainImpact: number;
    trafficImpact: number;
    timeImpact: number;
    seasonalImpact: number;
    historicalImpact: number;
  };
}

export class RoutePredictionService {
  private learningSystem: LearningSystem;
  private crossActivityLearning: CrossActivityLearningService;
  private contextualLearning: ContextualLearningService;
  private modelVersion = '1.0.0';

  constructor(
    learningSystem: LearningSystem,
    crossActivityLearning: CrossActivityLearningService,
    contextualLearning: ContextualLearningService
  ) {
    this.learningSystem = learningSystem;
    this.crossActivityLearning = crossActivityLearning;
    this.contextualLearning = contextualLearning;
  }

  async predictOptimalRoute(
    startPoint: [number, number],
    endPoint: [number, number],
    preferences: RoutePreferences,
    context: PredictionContext
  ): Promise<PredictionResult> {
    try {
      // Get historical patterns from learning system
      const historicalPatterns = await this.learningSystem.getPatterns({
        startPoint,
        endPoint,
        preferences,
        context: {
          weather: context.weather,
          terrain: context.terrain,
          traffic: context.traffic,
          time: {
            hour: context.timeOfDay,
            dayOfWeek: context.dayOfWeek,
            season: context.season
          }
        }
      });

      // Get cross-activity insights
      const crossActivityInsights = await this.crossActivityLearning.analyzePatterns({
        activityType: preferences.activityType,
        conditions: context.weather,
        terrain: context.terrain
      });

      // Get contextual adaptations
      const contextualAdaptations = await this.contextualLearning.getAdaptations({
        weather: context.weather,
        terrain: context.terrain,
        traffic: context.traffic,
        time: context.timeOfDay,
        season: context.season
      });

      // Combine all insights for prediction
      const prediction = await this.generatePrediction(
        startPoint,
        endPoint,
        preferences,
        context,
        historicalPatterns,
        crossActivityInsights,
        contextualAdaptations
      );

      return prediction;
    } catch (error) {
      console.error('Route prediction error:', error);
      throw new Error('Failed to predict optimal route');
    }
  }

  private async generatePrediction(
    startPoint: [number, number],
    endPoint: [number, number],
    preferences: RoutePreferences,
    context: PredictionContext,
    historicalPatterns: any,
    crossActivityInsights: any,
    contextualAdaptations: any
  ): Promise<PredictionResult> {
    // Calculate base route
    const baseRoute = await this.calculateBaseRoute(startPoint, endPoint, preferences);

    // Apply ML-based optimizations
    const optimizedRoute = await this.applyMLOptimizations(
      baseRoute,
      context,
      historicalPatterns,
      crossActivityInsights,
      contextualAdaptations
    );

    // Generate alternatives
    const alternatives = await this.generateAlternatives(optimizedRoute, context);

    // Calculate confidence and impact factors
    const { confidence, factors } = this.calculateConfidenceAndFactors(
      optimizedRoute,
      context,
      historicalPatterns,
      crossActivityInsights,
      contextualAdaptations
    );

    return {
      predictedRoute: optimizedRoute,
      confidence,
      alternativeRoutes: alternatives,
      factors
    };
  }

  private async calculateBaseRoute(
    startPoint: [number, number],
    endPoint: [number, number],
    preferences: RoutePreferences
  ): Promise<Route> {
    // Implement base route calculation
    // This would typically call your existing routing service
    return {
      id: `route-${Date.now()}`,
      name: 'Predicted Route',
      segments: [],
      preferences,
      createdAt: new Date(),
      updatedAt: new Date()
    };
  }

  private async applyMLOptimizations(
    baseRoute: Route,
    context: PredictionContext,
    historicalPatterns: any,
    crossActivityInsights: any,
    contextualAdaptations: any
  ): Promise<Route> {
    // Apply machine learning optimizations to the base route
    // This would use the insights from various learning systems
    return baseRoute;
  }

  private async generateAlternatives(
    optimizedRoute: Route,
    context: PredictionContext
  ): Promise<Route[]> {
    // Generate alternative routes based on the optimized route
    return [];
  }

  private calculateConfidenceAndFactors(
    route: Route,
    context: PredictionContext,
    historicalPatterns: any,
    crossActivityInsights: any,
    contextualAdaptations: any
  ): { confidence: number; factors: PredictionResult['factors'] } {
    // Calculate prediction confidence and impact factors
    return {
      confidence: 0.85,
      factors: {
        weatherImpact: 0.3,
        terrainImpact: 0.2,
        trafficImpact: 0.15,
        timeImpact: 0.15,
        seasonalImpact: 0.1,
        historicalImpact: 0.1
      }
    };
  }
} 